+++ app/Component/Request.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace App\Component;
+
+use Symfony\Component\HttpFoundation\Request as RequestBuilder;
+
+class Request
+{
+    private static $instance;
+
+    private function __construct()
+    {
+    }
+
+    public static function getInstance()
+    {
+        if (empty(self::$instance)) {
+            self::$instance = new RequestBuilder($_GET, $_POST, [], $_COOKIE, $_FILES, $_SERVER);
+        }
+        return self::$instance;
+    }
+}