+++ resources/views/index.blade.php
@@ -1,5 +1,5 @@
 @extends('_layouts.template')
 @section('title', 'index')
 @section('content')
-    body
+    hello
 @endsection